src/requirements.txt
MySQL-python==1.2.5
Django==1.11.1
django-allauth==0.32.0
django-vatno-validator==0.1.1
django-phonenumber-field==1.3.0
django-sslserver==0.19
python-magic==0.4.13
django-paypal==0.3.6
celery==4.0.2
django-celery-results==1.0.1
SpeechRecognition==3.7.1
django-timedeltatemplatefilter==0.1.2
